test(ui): fold SSE validation tests into validation.test.ts Remove sseEventValidation.test.ts; keep SSE cases next to other api/validation coverage in one file.
